renamed and cleaned linting exception TokenVisitor.swift [skip ci]
